creates a fs.Copy function, - re-uses most implementation from fs.Move
